+++ main/chopper/drive/DifferentialDriveSabertooth.cpp
@@ -7,17 +7,20 @@
 //     SetExpiration(950);
 // };
 
-void DifferentialDriveSabertooth::SetRampingValue(int ramping) {
+void DifferentialDriveSabertooth::SetRampingValue(int ramping)
+{
     // m_sabertoothDriver.setRamping((byte)(ramping));
     DifferentialDrive::SetRampingValue(ramping);
 }
 
-void DifferentialDriveSabertooth::SetDeadband(double deadband) {
+void DifferentialDriveSabertooth::SetDeadband(float deadband)
+{
     // m_sabertoothDriver.setDeadband((byte)(deadband*127));
     DifferentialDrive::SetDeadband(deadband);
 }
 
-void DifferentialDriveSabertooth::SetExpiration(uint64_t expirationTime) {
+void DifferentialDriveSabertooth::SetExpiration(uint64_t expirationTime)
+{
     // m_sabertoothDriver.setTimeout(expirationTime);
     DifferentialDrive::SetExpiration(expirationTime);
 };
